Change: Improve git based release id
Now cf-promises -T is not limited to working when you are on a checked
out branch. Now you get the commit that is checked out as the release id
even if you are checked out from a specific commit or tag.
(cherry picked from commit 4a47a88552455e0809bfd2874b1124bff67a92b4)