libutils/mustache.c: fix {{@}} key when value is not a primitive
Changelog: Mustache templates: Fix {{@}} key when value is not a
primitive. The old behavior, when iterating across a map or array of
maps, was to abort if the key was requested with {{@}}. The new
behavior is to always replace {{@}} with either the key name or the
iteration position in the array. An error is printed if {{@}} is used
outside of a Mustache iteration section.
(cherry picked from commit 30a146a67a27aea591112e821eee7b8085ca1ccb)