Redmine#7912: Fix classes being set because of hash collision in the implementation.
The implementation was using a Red-Black Tree as a global class
table. Unfortunately the key was not the full class name, but just a
hash of that. As a result, when two classes had the same hash, only one
had to be set for both to appear as set.

The fix removes the Red-Black Tree from the global class table, and
replaces with a hash table, which handles collisions natively.

Changelog: Title
(cherry picked from commit 0d930fbb16b4bf56b141155c382d9be8eca273d1)