Reverse order of binary lookup in testall.
It's actually the last path that ends up being used, not the first,
and it is this order of priority we want. This was motivated by trying
to run acceptance tests in a directory not named core, but core was
present next to it.
(cherry picked from commit c3ef07e48a7bb7926eff50d42d4819fbcefa1117)

Conflicts:

	tests/acceptance/testall