CFE-1803: storage: Properly initialize the list of current mounts
In commit 35523330, the sense of this check was inadvertently changed
from "If the list is empty, and LoadMountInfo fails, report an error" to
"If the list is non-empty, and LoadMountInfo fails, report an error".
This caused LoadMountInfo to never get called, so cf-agent assumed that
nothing was mounted and therefore that nothing needed to be unmounted.

Changelog: title
(cherry picked from commit c6bd79603a2b58a2ede99327176964156face6a6)