CFE-2506: Fix crash on Solaris when ps ucb variant is not available.
The correct approach when it is not available is to make the list of
ps information empty, not to delete it. This is already what happens
if the command can't be executed, but if executed with the wrong
arguments (ucb arguments for a non-ucb ps variant), the list would be
deleted.

Changelog: Title
Signed-off-by: Kristian Amlie <kristian.amlie@cfengine.com>
(cherry picked from commit 3da918718e95de4a1fefb4d5d02c31cf97916d78)