Slight adjustment to file_lib safe_open truncation test.
Actually it's the target that needs to be owned by root, and the
symlink should be owned by the user (which it is by default). Not
quite sure how it worked before, but with the latest change to allow
opening of symlinks that are either owned by root, or by the currently
running user, it needs to be like this. This still trigger the
intended unsafe path, since the target is sensitive and the symlink is
not.

Signed-off-by: Kristian Amlie <kristian.amlie@cfengine.com>
(cherry picked from commit d2054cb5ab851fbb1ce6214aca4965773cd568d8)