Added special behavior for bundles named __main__
If the bundle is defined in the entry policy it will be defined as main.
If the bundle is defined elsewhere, it will be removed.
This makes it easy to make importable library policy which can also
be executed directly.

Changelog: Commit
Signed-off-by: Ole Herman Schumacher Elgesem <ole.elgesem@northern.tech>
(cherry picked from commit 7bfd310b269a31a2b6740ff407fd16d36696758d)