Fix the argument for the '-b' option of cf-runagent to really be optional
The '--help' output says it is an optional argument, and the code
says 'optional_argument' as well. But for this to work with
'getopt_long()' two colons have to be used after the short option
character.

(cherry picked from commit 0e9fb8cd373ad0d55a6dc9525ec14626967e338f)