Respect windows newlines when splitting data in PipeReadData()
This needs to do different things on Windows and UNIX-like systems.

Add a function to split string on newlines to do so because it is
not trivial because the "\r\n" 2-byte sequence
represents a newline on Windows.

Ticket: ENT-3716
Changelog: None
(cherry picked from commit b6be8aa084d9413549a4d62fd6619500da09e5c9)