Add a function to get a quoted path (if needed)
Sometimes we need to quote a path, e.g. when passing it as an
argument to command execution in shell. However, the unqoted path
is much nicer in logs and other output if it doesn't need to be
quoted. So let's be nice on ourselves and the users and only
quote when we are not 100% sure quoting is not necessary.

Ticket: ENT-4432
Changelog: None
(cherry picked from commit 29137befbf4db317453b56c0afae55eabce48d92)