Quote the package module path when passing it to the interpreter
If the interpreter for a package module is specified, the package
module itself is passed as an argument to the interpreter (a
script path). If it contains spaces (or some other special
characters) it needs to be quoted.

Ticket: ENT-4432
Changelog: None
(cherry picked from commit 1b78077e43f82a18a7822dab569db53d94aeb3b2)