Fix potentially unitialized 'script_exec_opts' of a package module
If a package module has an interpreter specified, it can either
be just a path to the interpreter executable or it can also
provide options for the intepreter. If there are no such options,
'package_module->script_exec_opts' has to be NULL because we
later rely on that.

Changelog: None
(cherry picked from commit 74e3dd61e9081998dea11da15039b660616ecb02)