Augments files can now handle class expressions by appending '::'
A condition in an augments file is treated as a class expression
if it ends in `::`. Otherwise it is treated as a regular
expression.

Previously there was a bug, where the conditions in augments files
were treated as both class expressions and regular expressions.
This was ambiguous, and problematic since something like
`cfengine.nginx.enabled` as a regex would match a class called
`cfengine_nginx_enabled`, while as a class expression would match
`cfengine && nginx && enabled`.

This was fixed in 3.12.1 by treating it as a regex only.
(Documented and intended behavior).

Some users still wanted a way to use class expression, so the `::`
suffix solves the problem. For the example above, you can use this
in the augments file:

`cfengine.nginx.enabled::`

Since classes cannot end with `:`, this shouldn't be a problem,
it is not useful to have regular expressions ending with `::` in
this context.

Original commit made by @basvandervlies(bas.vandervlies@surfsara.nl).
Reworked by @olehermanse(ole@northern.tech).

Changelog: Commit
Ticket: CFE-2954
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit 097a6271fd24647540a7487dc19846d4df3f6297)