Update 3rdparty/peg and complete it
The libpromises/math.pc file which provides a parser for parsing
math expressions in the 'eval()' function calls was generated by
a tool called 'peg' in version 0.1.15. However, in our sources,
version 0.1.13 was committed and using it produced lots of
unrelated changes in the math.pc file.

Moreover, the Makefile from the project was missing so it was
impossible to just build it using 'make' (probably because 'git
add *' omits it as an ignored file and 'git add -f' has to be
used).

Also, the http://piumarta.com site is now gone so I had to use
https://web.archive.org/web/20160814190942/http://piumarta.com/software/peg
to get the sources.

Ticket: CFE-2762
Changelog: None
(cherry picked from commit 25c4f34ada5dccf2fbecda3df5af873046f1c03b)