Remember LMDB file path in MDB_env and use it in log messages
When we open an LMDB DB we get an MDB_env struct which we then
use to work with the DB. This struct has an option to store some
extra data (or a pointer to some data, to be more precise) which
gives us a chance to remember which DB file this MDB_env struct
was created for and then we can this information in log messages.

Technically, the MDB_env struct can be used to work with multiple
DBs, but that's not how we use it.

Ticket: CFE-2469
Changelog: Local DB errors are now logged with the particular DB file path
(cherry picked from commit f88d1a59b19c2f58ecae55bd82563e87b84f1408)