Only use AI_NUMERICHOST for what looks like an IPv4 address
Using AI_NUMERICHOST for getaddrinfo() is an optimization that
allows us to avoid potential long DNS queries when doing bind()
in cf-server. However, it cannot be used if the bind address is a
host name and not an IP address.

This change makes it being used only in cases where the bind
address looks like an IPv4 address. If it doesn't, getaddrinfo()
will take care of resolving things. Please note that this flag is
just a hint for getaddrinfo(), not using it never changes the
results (using it potentiall does).

Ticket: CFE-3190
Changelog: Allow using hostname in bindtoaddress
(cherry picked from commit 6f6f688ad8484d73d910741c2490b1e4cefb8513)