Try to kill the whole process group in cf-execd when agent_expireafter
'exec_command' to spawn the agent process from cf-execd is
actually executed in a shell. So if 'agent_expireafter' triggers,
we need to kill the agent process not just the shell it is
running in.

IOW, we need to find the process that has the shell as its parent
process. And if we find a process group of with the same PID as
such process, we should kill the whole process group, not just
the process itself.

Ticket: CFE-2664
Changelog: cf-execd now sends SIGKILL to the agent process in case of agent_expireafter
(cherry picked from commit 31ba2e6fb41112a794b895d906c325974ff3417c)