Do not send signals to a dead process in ProcessSignalTerminate
We try sending SIGINT, SIGTERM and SIGKILL to the target process
in ProcessSignalTerminate(). If one of the SIGNALS work, we
should stop and not try to send the other signals to the process
that no longer exists. That only results in noise in logs.

(cherry picked from commit 7bf08023a175d50d6fe492377bbe8f77297465b5)