Always set $(sys.libdir) and $(sys.local_libdir)
These no longer depend on version so they should be set in all
cases not just when the CFEngine version is known and parsed.

The code should not be in DiscoverVersion(), but that's out of
scope of the changes done here.

(cherry picked from commit 60d4dae1a5457740b01ab78b114336a295793b7a)