Allow cf-serverd to connect to hub for call-collect
When call-collect (client initiated reporting) is used, the
cf-serverd process needs to connect to the hub. This is not
allowed by default for constrained SELinux domains so we need to
add a rule for that.

Ticket: ENT-5415
Changelog: Client initiated reporting was fixed on RHEL 8.1
(cherry picked from commit 88375879c2c9b2df794a4c209dc9c4ec463b17c0)