Add multi-key/multi-host support to cf-keycrypt
Encrypting secrets for hosts individually is error-prone and
results in many files. A better approach is to encrypt the
payload for multiple hosts (keys) at once and store the result in
a file the format of which allows the individual hosts to decrypt
the payload back using their keys.

This commit introduces a new 'Encrypted-for' header for the
encrypted file format which tells which keys the payload was
encrypted for and also what is the ordering of the segments
containing the encrypted symmetric key. Thus the format now looks
as follows:

  ------------
  | Headers  |
  ------------
  | AES IV   |
  ------------
  | AES key  |
  ------------
  | AES key  |
  ------------
  | ...      |
  ------------
  | AES key  |
  ------------
  | data     |
  ------------

Where each 'AES key' block is the same randomly generated AES key
encrypted by different RSA (host) key. The host then needs to
identify which 'AES key' block to decrypt and use as the key to
decrypt the payload.

Ticket: CFE-3271
Changelog: None
(cherry picked from commit 19d007eb49071caa8ec8ad16f0b176e22d214d9a)