Properly detect end of the headers section in 'cf-secret print-headers'
Attempt to read the next header after the last one fails because
the ParseHeader() function just reads until it hits the maximum
limit of characters to read or the key-value delimiter (':'). The
check for the end of the headers section of the file needs to be
done outside of this function, just like it is done in
ParseHeaders() which is part of the decryption process.

Ticket: CFE-3368
Changelog: None
(cherry picked from commit 0a94bfc9c494698ef1a20bfa3d5bacfefc91e74f)