Log a NOTICE-level message when force-writing critical section lock
This is something that should ideally never happen and if it
does, it should definitely be logged. At the same time, it is not
exactly an error in the process that logs it because it is just
dealing with some trash left behind by some other process. Still
a message worth noticing, though.

Ticket: CFE-3361
Changelog: None
(cherry picked from commit dabd47543f1733f94697b4008449ea24034a9942)