Don't try to backup files with an empty name
When creating new files (for example with a copy_from when the destination
does not exist), if backups are enabled, the agent would try to backup the
previous state of the (non-existent) destination. It makes the promise
fail and logs errors about the backup of an empty file name.

Changelog: Don't fail on new file creation when backups are enabled
Ticket: CFE-3640
(cherry picked from commit e97d5fb4f2fcf9dcd5a302defb03187827ca4257)