Add a special SELinux type and domain for Mission Portal action scripts
The notification scripts that can be hooked up to alerts are
supposed to be able to do basically anything on the system. Thus
we need to allow that, but we shouldn't enable our httpd and PHP
do anything on the system.

Ticket: ENT-8029
Changelog: None
(cherry picked from commit 3f25d27c3f2b26668c7461124da8d4a68d6bf098)