Added config to EvalContext
Since this was already in a comment that somebody needed it
in the past, and I need it now, I went ahead and added it.

Had to adjust diff slightly because of changes in surrounding
code (restrict keys functions).

Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit fe1f0928cb71568097b59f87a7996f0bee0ec9fd)