Fixed inconsistencies with methods promises and missing bundles
Previously, methods promises acted differently depending on if
you specify the bundle with usebundle or in the promiser string.

This change removes the inconsistent checking which was only
happening on promises with usebundle. It also ensures that
the agent will abort if trying to use an undefined bundle,
regardless of whether it was using promiser string or usebundle.

The checking that was removed did allow you to detect some
missing bundles early (in cf-promises), but it was also not
working correctly, especially in combination with the bug
we fixed for ignore_missing_bundles.

If we want to reintroduce this logic, we have to do it in
a better way, we need to determine which bundles are
"reachable", by starting at the bundlesequence and
traversing methods promises, while respecting if
attributes and class guards. That would be useful, but
quite a bit more complex.

This change, combined with the fix for ignore_missing_bundles
and ignore_missing_inputs, allow you to use inputs and bundles
conditionally (like we do for the enterprise federation policy)
and still have the agent abort in the correct situations
(when a bundle is actually missing and you haven't enabled
ignore_missing_bundles). The downside is that there are some
potential situations where undefined bundles would be detected
earlier previously (both correctly and incorrectly).

Changelog: Commit
Ticket: ENT-8430
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>
(cherry picked from commit a7e1eb6a081dee4d237392e9ca1d058d3b940fe9)