Fixed misleading error message when adding CMDB classes
The function that adds CMDB classes to the context, checks if the class
is already there. If this is the case, it skips adding the class and
returns false. The code using this function interpreted false as an
error, which is not the case.

Ticket: ENT-10336
Changelog: None
Signed-off-by: Lars Erik Wik <lars.erik.wik@northern.tech>
(cherry picked from commit 0623d7d5b018f4ea5d853401f7b2e9dd37c1dac3)