Drop VerifyThatDatabaseIsNotCorrupt() on locks DB
This function totally doesn't do what its name says. It only
checks if the DB was modified in the current boot and if not, it
restores it from the latest backup. Which is done and the end of
every agent run. So this function effectively reverts the locks
DB to the state of the last agent run on every boot dropping
significant data like daemon locks.

Ticket: CFE-3982
Changelog: cf_lock.lmdb is no longer restored from backup on
           every boot
(cherry picked from commit 4ec4e28748e71c642dea47b1222e1e060b197016)