Hack (fixme) to enable cf-serverd to bind on ipv4 address
As-is, it seems that querying for AF_UNSPEC on OpenBSD only returns AF_INET6 which also apparently does not bind to AF_INET(ipv4) as it does on other platforms.

For now this is hard-coded. Will work on an ifdef for OpenBSD or some code to workaround this issue.

Ticket: none
Changelog: title (when there is a real fix)