Legacy package methods refactored slightly to be clear about available packages
Before the code used the variable name/logging messsage "matches" where it really meant "available" packages.

A verbose log has been added to acknowledge that a package will be installed even if it is not in the available packages.
This is because most package_method bodies in masterfiles/lib/packages.cf do not distinguish between installed or available packages.

This change aims to clarify in the code and logging what is happening and answer the comment: "why do we schedule a 'not matched' operation?".

We schedule it because that is the legacy behavior and changing it would involve quite a bit of refactoring.

Ticket: ENT-13578
Changelog: none